pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP server with JWT auth, weather and crypto tools"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]",
    "httpx",
    "pyjwt",
    "uvicorn",
    "pydantic-settings",
    "openai",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "agent"]
